test(coverage): add sweep suites for engine, dag and tools; fix bugs they exposed
New unit test binaries (wired into unit-tests, -asan, run_all_suite.sh and
the coverage build/report):
- test_coverage_sweep: query engine matrix/truncation/mutation, wire/TSIG/EDNS,
  zone/config/tinydns parsers, snapshot+catalog, XFR, UPDATE, RRL/ECS/dnstap
- test_coverage_sweep_dag: dag option tokens and print_response display matrix
- test_coverage_sweep_net: dag against a fake UDP/TCP/DoT/DoH server
  (trace/nssearch, AXFR/IXFR, UPDATE, SIG(0) key loaders, --replay inputs)
- test_coverage_sweep_tools: karicheck lint rules, karictl control protocol

Fixes:
- dns_snapshot_rcu: compare catalog member groups using the right indices
- dns_catalog_zone: require a single-label unique-id in member names
- dns_config_parser: avoid int overflow in buffer-size suffix multiplication
- karicheck: reject non-regular files (ftell() on a directory overflowed)
- test_dag_tools: avoid signed shift UB in zigzag encoding

coverage_gate.pl now includes the fault-injection and sweep binaries, so the
gate matches the HTML report.

+++ tests/test_dag_tools.c
@@ -4422,7 +4422,7 @@ static void test_dag_tools_feature_case_202(void) {
 static void test_dag_tools_feature_case_203(void) {
     printf("[TEST] DAG Tools: Replay Protobuf zigzag 32-bit encoding...\n");
     int32_t n = -1;
-    uint32_t zz = (uint32_t)((n << 1) ^ (n >> 31));
+    uint32_t zz = ((uint32_t)n << 1) ^ (uint32_t)(n >> 31);   /* shift the unsigned value: n << 1 on a negative int is UB */
     assert(zz == 1);
 }
